updates files and splits formal site
