+++ .gitignore
@@ -29,6 +29,7 @@ yarn-error.log*
 .env.development.local
 .env.test.local
 .env.production.local
+secrets.json
 
 # node
 package-lock.json
