+++ components/Shortcut/styles.module.scss
