+++ components/StickyFooter/styles.module.scss
@@ -21,7 +21,7 @@
 
 .taskWindow {
   display: flex;
-  justify-content: start;
+  justify-content: flex-start;
   flex-direction: row;
   align-items: center;
