+++ package.json
@@ -23,9 +23,12 @@
     "sass": "^1.39.0"
   },
   "devDependencies": {
+    "@types/node": "^20.2.5",
+    "@types/react": "^18.2.7",
     "eslint": "7.32.0",
     "eslint-config-next": "^13.4.4",
-    "file-loader": "^6.2.0"
+    "file-loader": "^6.2.0",
+    "typescript": "^5.0.4"
   },
   "engines": {
     "node": "18.x",
