+++ pages/_app.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+import type { AppProps } from 'next/app'
+import '../styles/globals.scss'
+
+export default function App({ Component, pageProps }: AppProps) {
+  return <Component {...pageProps} />
+}
