+++ pages/home/index.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import styles from './styles.module.scss'
+
+const Home = () => {
+  return (
+    <>
+      <div>hello world</div>
+    </>
+  )
+}
+
+export default Home
